[chore] html 동적 생성 및 insert 작업은 component 내 js에서 처리하도록 코드 이동
